# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prometheus API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/projects": {
      "get": {
        "summary": "List Projects",
        "operationId": "list_projects_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Project",
        "operationId": "create_project_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}": {
      "get": {
        "summary": "Get Project",
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/cad-imports": {
      "post": {
        "summary": "Import Cad",
        "operationId": "import_cad_projects__project_id__cad_imports_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "fixture",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Fixture"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_cad_projects__project_id__cad_imports_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/component-research": {
      "post": {
        "summary": "Research Component",
        "operationId": "research_component_component_research_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/component-packages/{package_id}": {
      "get": {
        "summary": "Get Package",
        "operationId": "get_package_component_packages__package_id__get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Package Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/component-packages/{package_id}/confirm": {
      "post": {
        "summary": "Confirm Package",
        "operationId": "confirm_package_component_packages__package_id__confirm_post",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Package Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/connections": {
      "post": {
        "summary": "Create Connection",
        "operationId": "create_connection_projects__project_id__connections_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/scenarios": {
      "post": {
        "summary": "Create Scenario",
        "operationId": "create_scenario_projects__project_id__scenarios_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scenarios/{scenario_id}/compile": {
      "post": {
        "summary": "Compile Scenario",
        "operationId": "compile_scenario_scenarios__scenario_id__compile_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scenarios/{scenario_id}/runs": {
      "post": {
        "summary": "Run Checks",
        "operationId": "run_checks_scenarios__scenario_id__runs_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/findings": {
      "get": {
        "summary": "Get Findings",
        "operationId": "get_findings_runs__run_id__findings_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_import_cad_projects__project_id__cad_imports_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "type": "object",
        "title": "Body_import_cad_projects__project_id__cad_imports_post"
      },
      "ConnectionCreate": {
        "properties": {
          "source_part": {
            "type": "string",
            "title": "Source Part"
          },
          "target_part": {
            "type": "string",
            "title": "Target Part"
          },
          "connection_type": {
            "type": "string",
            "enum": [
              "fixed",
              "revolute",
              "sliding"
            ],
            "title": "Connection Type",
            "default": "revolute"
          },
          "axis": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Axis",
            "default": [
              0,
              0,
              1
            ]
          },
          "limits_deg": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Limits Deg",
            "default": [
              0,
              90
            ]
          }
        },
        "type": "object",
        "required": [
          "source_part",
          "target_part"
        ],
        "title": "ConnectionCreate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "unit_system": {
            "type": "string",
            "enum": [
              "SI",
              "US"
            ],
            "title": "Unit System",
            "default": "SI"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreate"
      },
...
      "ResearchRequest": {
        "properties": {
          "manufacturer": {
            "type": "string",
            "title": "Manufacturer",
            "default": "Prometheus Fixture Works"
          },
          "part_number": {
            "type": "string",
            "title": "Part Number",
            "default": "PM-36-GM"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "type": "object",
        "title": "ResearchRequest"
      },
...
      "ScenarioCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": "Motor arm duty cycle"
          },
          "natural_language_description": {
            "type": "string",
            "title": "Natural Language Description"
          },
          "definition": {
            "$ref": "#/components/schemas/ScenarioDefinition"
          }
        },
        "type": "object",
        "required": [
          "natural_language_description",
          "definition"
        ],
        "title": "ScenarioCreate"
      },
      "ScenarioDefinition": {
        "properties": {
          "payload_kg": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Payload Kg"
          },
          "arm_length_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Arm Length M"
          },
          "rotation_deg": {
            "type": "number",
            "maximum": 360.0,
            "exclusiveMinimum": 0.0,
            "title": "Rotation Deg"
          },
          "movement_s": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Movement S"
          },
          "hold_s": {
            "type": "number",
            "minimum": 0.0,
            "title": "Hold S"
          },
          "cycle_s": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cycle S"
          },
          "ambient_c": {
            "type": "number",
            "maximum": 150.0,
            "minimum": -40.0,
            "title": "Ambient C"
          },
          "safety_factor": {
            "type": "number",
            "minimum": 1.0,
            "title": "Safety Factor",
            "default": 1.2
          }
        },
        "type": "object",
        "required": [
          "payload_kg",
          "arm_length_m",
          "rotation_deg",
          "movement_s",
          "hold_s",
          "cycle_s",
          "ambient_c"
        ],
        "title": "ScenarioDefinition"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}